Refactor: 디렉토리 구조 변경
디렉토리 구조를 크게 domain과 global로 나누었습니다.
