--- src/main/java/com/service/dida/comment/Comment.java
@@ -1,37 +1,0 @@
-package com.service.dida.comment;
-
-import com.service.dida.common.BaseEntity;
-import com.service.dida.post.Post;
-import com.service.dida.user.User;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "comment")
-public class Comment extends BaseEntity {
-    @Id
-    @Column(name = "comment_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
-
-    @Column(name = "content", nullable = false, length = 200)
-    private String content;
-
-    @Column(name = "deleted", nullable = false, columnDefinition = "boolean default false")
-    private boolean deleted;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-}
