--- src/main/java/com/service/dida/config/exception/BaseException.java
@@ -1,15 +1,0 @@
-package com.service.dida.config.exception;
-
-import lombok.Getter;
-
-@Getter
-public class BaseException extends RuntimeException {
-
-    private final String errorCode;
-    private final String message;
-
-    public BaseException(ErrorCode code) {
-        errorCode = code.getErrorCode();
-        message = code.getMessage();
-    }
-}
