--- src/main/java/com/service/dida/config/exception/ErrorCode.java
@@ -1,7 +1,0 @@
-package com.service.dida.config.exception;
-
-public interface ErrorCode {
-    String getErrorCode();
-
-    String getMessage();
-}
