+++ src/main/java/com/service/dida/domain/market/MarketRepository.java
@@ -1,0 +1,6 @@
+package com.service.dida.domain.market;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MarketRepository extends JpaRepository<Market, Long> {
+}
