+++ src/main/java/com/service/dida/domain/user/Role.java
@@ -1,0 +1,13 @@
+package com.service.dida.domain.user;
+
+public enum Role {
+
+    GENERAL("일반 회원"),
+    MANAGER("관리자");
+
+    private final String name;
+
+    Role(String name) {
+        this.name = name;
+    }
+}
