+++ src/main/java/com/service/dida/domain/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.service.dida.domain.user.repository;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.service.dida.domain.user.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUserId(Long userId);
+
+}
