+++ src/main/java/com/service/dida/global/config/exception/ErrorCode.java
@@ -1,0 +1,7 @@
+package com.service.dida.global.config.exception;
+
+public interface ErrorCode {
+    String getErrorCode();
+
+    String getMessage();
+}
