--- src/main/java/com/service/dida/like/Like.java
@@ -1,33 +1,0 @@
-package com.service.dida.like;
-
-import com.service.dida.nft.Nft;
-import com.service.dida.user.User;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "nft_like")
-public class Like {
-    @Id
-    @Column(name = "like_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long likeId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "nft_id")
-    private Nft nft;
-
-    @Column(name = "status", nullable = false, columnDefinition = "boolean default true")
-    private boolean status;
-}
