--- src/main/java/com/service/dida/market/MarketRepository.java
@@ -1,6 +1,0 @@
-package com.service.dida.market;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MarketRepository extends JpaRepository<Market, Long> {
-}
