--- src/main/java/com/service/dida/market/dto/GetMainPageWithoutSoldOut.java
@@ -1,15 +1,0 @@
-package com.service.dida.market.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class GetMainPageWithoutSoldOut {
-    List<GetHotItem> getHotItems;
-    List<GetHotSeller> getHotSellers;
-    List<GetRecentNft> getRecentNfts;
-    List<GetHotUser> getHotUsers;
-}
